backend/requirements.txt
Django>=5.0.6
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

